Optimise several data types and code paths. Use a new "above" key for tilesets instead of using an attribute. Move external repositories outside to root instead of src.

+++ src/engine/include/engine/time.h
@@ -15,10 +15,10 @@ public:
         Minute,
     } Type;
 
-    static void parseString(const std::string& format, std::unordered_map<Type, int>& out);
+    static void parseString(const std::string& format, std::unordered_map<Type, int8_t>& out);
 
 private:
-    static void parseBlock(const std::string& block, std::unordered_map<Type, int>& out);
+    static void parseBlock(const std::string& block, std::unordered_map<Type, int8_t>& out);
 
 };
 }
